- 기존 클래스들의 의존성을 낮추기 위해서 불필요한 클래스를   제거하였습니다. - 기존의 축약해서 사용헀던 변수명을 정확하게 표기하였습니다.
    1. Game
	- main 메소드가 있는 클래스 입니다.
	- game클래스가 생성될 때, 생성자에서 inputMonsterCount,
	  inputGameCount메소드를 실행해서 출력 메세지와 함께
	  monsterCount, gameCount 변수에 값을 입력받아 저장하게
	  하였습니다.
    2. Stadium
	- 스타디움 클래스는 경기가 일어나는 공간이라고 생각하고, 경기를
	  진행하는 메소드와, 경기결과를 출력하는 메소드를 가지고
	  있습니다.
	- 스타디움이 생성되면, 생성자에서 playStart라는 메소드를
	  호출하여 경기를 시작합니다.
	- 경기의 결과는 main 함수에서 직접 stadium 클래스 내부에 있는
	  결과를 불러와서 확인하게 하였습니다.
