--- moster/src/main/java/MonsterRace/OutputGameResult.java
@@ -1,28 +1,0 @@
-package MonsterRace;
-
-import java.util.List;
-
-public class OutputGameResult {
-
-    public OutputGameResult(GameInfo gameInfo) {
-        printResult(gameInfo);
-    }
-
-
-    public void printResult(GameInfo gameInfo) {
-        List<Monster> playerList = gameInfo.getPlayerList();
-        playerList.stream()
-                .forEach(monster -> {
-                    System.out.println(makeDash(monster));
-                });
-    }
-
-    public String makeDash(Monster monster) {
-        int totalMove = monster.getTotMove();
-        StringBuilder sb = new StringBuilder();
-        for(int i = 0; i < totalMove; i++){
-            sb.append("-");
-        }
-        return sb.toString();
-    }
-}
