+++ moster/src/main/java/MonsterRace/Stadium.java
@@ -5,38 +5,51 @@ import java.util.List;
 import java.util.Random;
 
 public class Stadium {
+    List<Monster> playerList = new ArrayList<>();
 
-    public Stadium(GameInfo gameInfo) {
-        playGame(gameInfo);
+    public Stadium(int monsterCount, int gameCount) {
+        playGame(monsterCount,gameCount);
     }
 
-    private void playGame(GameInfo gameInfo) {
-        initStadium(gameInfo);
-        gameInfo.getPlayerList().stream()
+    private void playGame(int monsterCount,int gameCount) {
+        makePlayerList(monsterCount);
+        playerList.stream()
                 .forEach(monster ->{
-                    movingMonster(monster,gameInfo.getGameCnt());
+                    movingMonster(monster, gameCount);
                 });
-
     }
 
-    private void initStadium(GameInfo gameInfo) {
-        int monsterCnt = gameInfo.getMonsterCnt();
-        List<Monster> tempPlayerList = new ArrayList<>();
-        for(int i = 0; i < monsterCnt; i++){
-            tempPlayerList.add(new Monster(i));
+    private void makePlayerList(int monsterCount) {
+        for(int i = 0; i < monsterCount; i++){
+            playerList.add(new Monster(i));
         }
-        gameInfo.setPlayerList(tempPlayerList);
     }
 
     private void movingMonster(Monster monster, int gameCnt) {
         Random random = new Random();
         int ranNum = random.nextInt(10);
-        if(ranNum > 4){
-            monster.setTotMove(monster.getTotMove()+1);
+        if (ranNum > 4) {
+            monster.setTotalMove(monster.getTotalMove()+1);
         }
-        if(gameCnt > 0){
+        if (gameCnt > 0) {
             movingMonster(monster,gameCnt-1);
         }
+    }
+
+    public void printResult() {
+        playerList.stream()
+                .forEach(monster -> {
+                    System.out.println(makeDash(monster));
+                });
+    }
+
+    public String makeDash(Monster monster) {
+        int totalMove = monster.getTotalMove();
+        StringBuilder sb = new StringBuilder();
+        for(int i = 0; i < totalMove; i++){
+            sb.append("-");
+        }
+        return sb.toString();
     }
 
 }
